Use_OpenAI/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
openai==1.55.0
pydantic==2.10.0
pydantic_core==2.27.0
python-dotenv==1.0.1
sniffio==1.3.1
tenacity==9.0.0
tqdm==4.67.0
typing_extensions==4.12.2
